+++ src/main/java/com/backend/girnartour/RequestDTOs/SalesReceiptRequest.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 
 @Data
 @AllArgsConstructor
@@ -12,7 +14,7 @@ public class SalesReceiptRequest {
 
     private String date;
 
-    private Double amountReceived;
+    private BigDecimal amountReceived;
 
     private String receiptType;
